Show mappings from source vocabularies to standard

Show mappings from source vocabularies to standard

show_mappings(
  candidate_codelist,
  source_vocabularies = c("ATC", "ICD10CM", "ICD10PCS", "ICD9CM", "ICD9Proc", "LOINC",
    "OPCS4", "Read", "RxNorm", "RxNorm Extension", "SNOMED"),
  db,
  vocabulary_database_schema
)

Arguments

candidate_codelist

Dataframe

source_vocabularies

Character vector

concept

Dataframe

concept_relationship

Dataframe

Value

Dataframe

Examples

# note, Eunomia, which is used for the example below, does not include a full set of vocabularies. The full set can be downloaded from https://athena.ohdsi.org
untar(xzfile(system.file("sqlite", "cdm.tar.xz", package = "Eunomia"), open = "rb"),
      exdir =  tempdir())
db <- dbConnect(RSQLite::SQLite(), paste0(tempdir(),"\\cdm.sqlite"))
#> Error in dbConnect(RSQLite::SQLite(), paste0(tempdir(), "\\cdm.sqlite")): could not find function "dbConnect"
asthma_codes<-get_candidate_codes(keywords="asthma",
                  db=db,
                  vocabulary_database_schema = "main")
#> Error in tbl(db, sql(paste0("SELECT * FROM ", vocabulary_database_schema,     ".concept"))): could not find function "tbl"
show_mappings(candidate_codelist= asthma_codes,
                  db=db,
                  vocabulary_database_schema = "main")
#> Error in isTRUE(lhs): object 'asthma_codes' not found