Stratify a codelist by route category.

Stratify a codelist by route category.

Usage

stratifyByRouteCategory(x, cdm, keepOriginal = FALSE)

Arguments

x

A codelist.

cdm

A cdm reference via CDMConnector.

keepOriginal

Whether to keep the original codelist and append the stratify (if TRUE) or just return the stratified codelist (if FALSE).

Value

The codelist with the required stratifications, as different elements of the list.

Examples

# \donttest{
library(CodelistGenerator)
cdm <- mockVocabRef()
#> Warning: There are observation period end dates after the current date: 2025-04-11
#> ℹ The latest max observation period end date found is 2025-12-31
codes <- list("concepts" = c(20,21))
new_codes <- stratifyByRouteCategory(x = codes,
                                     cdm = cdm,
                                     keepOriginal = TRUE)
#> Warning: ! `codelist` casted to integers.
#> Warning: ! `codelist` casted to integers.
new_codes
#> 
#> ── 3 codelists ─────────────────────────────────────────────────────────────────
#> 
#> - concepts (2 codes)
#> - concepts_topical (1 codes)
#> - concepts_transmucosal_nasal (1 codes)
# }