Generate candidate codelist for the OMOP CDM

This function generates a set of codes that can be considered for creating a phenotype using the OMOP CDM.

Usage

getCandidateCodes(
  keywords,
  domains = "Condition",
  conceptClassId = NULL,
  standardConcept = "Standard",
  searchSynonyms = FALSE,
  searchSource = FALSE,
  fuzzyMatch = FALSE,
  maxDistanceCost = 0.1,
  exclude = NULL,
  includeDescendants = TRUE,
  includeAncestor = FALSE,
  verbose = FALSE,
  db,
  vocabularyDatabaseSchema
)

Arguments

keywords

Character vector of words to search for. Where more than one word is given (e.g. "knee osteoarthritis"), all combinations of those words should be identified positions (e.g. "osteoarthritis of knee") should be identified.

domains

Character vector with one or more of the OMOP CDM domain.

conceptClassId

Character vector with one or more concept class of the Concept

standardConcept

Character vector with one or more of "Standard", "Classification", and "Non-standard". These correspond to the flags used for the standard_concept field in the concept table of the cdm.

searchSynonyms

Either TRUE or FALSE. If TRUE the code will also search via the concept synonym table.

searchSource

Either TRUE or FALSE. If TRUE the code will also search via source concepts.

fuzzyMatch

Either TRUE or FALSE. If TRUE the fuzzy matching will be used, with approximate matches identified.

maxDistanceCost,

The maximum number/fraction of match cost (generalized Levenshtein distance) for fuzzy matching (see ??base::agrep for further details).

exclude

Character vector of words to identify concepts to exclude.

includeDescendants

Either TRUE or FALSE. If TRUE descendant concepts of identified concepts will be included in the candidate codelist.

includeAncestor

Either TRUE or FALSE. If TRUE the direct ancestor concepts of identified concepts will be included in the candidate codelist.

verbose

Either TRUE or FALSE. If TRUE, progress will be reported.

db

Database connection via DBI::dbConnect()

vocabularyDatabaseSchema

Name of database schema with vocab tables

Value

tibble

Examples

if (FALSE) {
library(DBI)
library(CodelistGenerator)
db <- DBI::dbConnect(" Your database connection here ")
vocabularyDatabaseSchema <- " Your vocabulary schema here "
getCandidateCodes(
  keywords = "asthma",
  db = db,
  vocabularyDatabaseSchema = vocabularyDatabaseSchema
)
}