Get descendant codes for drug ingredients

Get descendant codes for drug ingredients

Usage

getDrugIngredientCodes(
  cdm,
  name = NULL,
  nameStyle = "{concept_code}_{concept_name}",
  doseForm = NULL,
  doseUnit = NULL,
  routeCategory = NULL,
  ingredientRange = c(1, Inf),
  type = "codelist"
)

Arguments

cdm

cdm_reference via CDMConnector

name

Names of ingredients of interest. For example, c("acetaminophen", "codeine"), would result in a list of length two with the descendant concepts for these two particular drug ingredients. Users can also specify the concept ID instead of the name (e.g., c(1125315, 42948451)) using a numeric vector.

nameStyle

Name style to apply to returned list. Can be one of "{concept_code}_{concept_name}", "{concept_code}", or "{concept_name}".

doseForm

Only descendants codes with the specified dose form will be returned. If NULL, descendant codes will be returned regardless of dose form.

doseUnit

Only descendants codes with the specified dose unit will be returned. If NULL, descendant codes will be returned regardless of dose unit

routeCategory

Only descendants codes with the specified route will be returned. If NULL, descendant codes will be returned regardless of route category.

ingredientRange

Used to restrict descendant codes to those associated with a specific number of ingredients. Must be a vector of length two with the first element the minimum number of ingredients allowed and the second the maximum. A value of c(2, 2) would restrict to only concepts associated with two ingredients.

type

Can be "codelist", "codelist_with_details", or "concept_set_expression"

Value

Concepts with their format based on the type argument.

Examples

# \donttest{
cdm <- mockVocabRef()
#> Warning: There are observation period end dates after the current date: 2025-03-27
#> ℹ The latest max observation period end date found is 2025-12-31
getDrugIngredientCodes(cdm = cdm, name = "Adalimumab",
                       nameStyle = "{concept_name}")
#> 
#> ── 1 codelist ──────────────────────────────────────────────────────────────────
#> 
#> - adalimumab (2 codes)
# }